Added logic for turning arrow

+++ FindWayHome/app/src/main/java/se/keyelementab/findwayhome/DirectionManager.kt
@@ -1,14 +1,10 @@
 package se.keyelementab.findwayhome
 
-import android.content.Context
-import android.location.Location
-import android.util.Log
-
 class DirectionManager {
     private val TAG = "DirectionManager"
 
     /**
-     * TODO
+     * Calculates how many degrees to turn the image to point to the destination.
      *
      * @param bearingToDestination the direction to the destination in degrees relating to true north
      * , i.e. not magnetic north, ranging from -180 to 180.
@@ -17,14 +13,8 @@ class DirectionManager {
      * @return the degrees to turn the image, ranging from -180 to 180.
      */
     fun degreesToTurnImage(bearingToDestination : Float,
-                           bearingCompass : Float) : Int {
-
-
+                           bearingCompass : Float) : Float {
         var degreesToTurn = 360f - bearingCompass + bearingToDestination
-
-   //     Log.d(TAG, "bearingToDestination " + bearingToDestination)
-     //   Log.d(TAG, "bearingCompass " + bearingCompass)
-//        Log.d(TAG, "degreesToTurn " + degreesToTurn)
 
         if (degreesToTurn <= -360) {
             degreesToTurn = degreesToTurn + 360
@@ -34,6 +24,6 @@ class DirectionManager {
             degreesToTurn = degreesToTurn - 360
         }
 
-        return degreesToTurn.toInt()
+        return degreesToTurn
     }
 }
